// src/sanity/sanity.types.ts
export type ReleaseLink = {
  _type: 'releaseLink'
  linkType?: 'Buy' | 'Stream' | 'Vinyl' | 'Other'
  url?: string
  label?: string
}
// ...
export type Release = {
  _id: string
  _type: 'release'
  _createdAt: string
  _updatedAt: string
  _rev: string
  name?: string
  slug?: Slug
  artists?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'artist'
  }>
  releaseDate?: string
  bandcampEmbedUrl?: string
  links?: Array<
    {
      _key: string
    } & ReleaseLink
  >
  buyLink?: string
  streamLink?: string
  releaseType?: 'LP' | 'EP' | 'Single'
}

export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}

export type AllSanitySchemaTypes =
  | SanityImagePaletteSwatch
  | SanityImagePalette
  | SanityImageDimensions
  | SanityImageHotspot
  | SanityImageCrop
  | SanityFileAsset
  | SanityImageAsset
  | SanityImageMetadata
  | Geopoint
  | SanityAssetSourceData
  | ReleaseLink
  | SubmissionSettings
  | Artist
  | Release
  | Slug
export declare const internalGroqTypeReferenceTo: unique symbol